using mixins strategy for API. better docs
