+++ README.md
@@ -20,9 +20,8 @@ About promises.
 ### Starting a chain
 There are three types of methods that can initiate a new chain:
 
-#### bind
-`bind( key<String>, modifier<Array> )`
-
+**bind**  
+`bind( key<String>, modifier<Array> )`  
 Listen to a particular key combination and fire the chain every time such
 shortcut is triggered.
 
@@ -32,9 +31,8 @@ z.bind('t', ['Cmd', 'Shift']).then(function(){
 });
 ```
 
-#### listen
-`listen( event<String> )`
-
+**listen**  
+`listen( event<String> )`  
 Listen to a particular event and fire the chain every time this event occurs.
 
 ```node
@@ -43,9 +41,8 @@ z.listen('window_created').then(function(){
 });
 ```
 
-#### api
-`api()`
-
+**api**  
+`api()`  
 Manually trigger a displsable chain of actions.
 
 ```node
@@ -58,9 +55,8 @@ When a new chain is initialised, ypu can use the API to interact with Windows,
 Apps, Screens and Ohter stuff.
 
 ### Window
-#### getFocusedWindow
-`getFocusedWindow() ⇒ object`
-
+**getFocusedWindow**  
+`getFocusedWindow() ⇒ object`  
 Return an object with the id of the focused window.
 
 ```node
@@ -69,31 +65,8 @@ z.bind('t', ['Cmd', 'Shift']).getFocusedWindow().then(function(window){
 });
 ```
 
-#### getVisibleWindows
-`getVisibleWindows() ⇒ array`
-
-Return an array containing an object for every visible window.
-
-```node
-z.bind('t', ['Cmd', 'Shift']).getVisibleWindows().then(function(windows){
-  window.forEach(console.log);
-});
-```
-
-#### getAllWindows
-`getAllWindows() ⇒ array`
-
-Return an array containing an object for all the windows.
-
-```node
-z.bind('t', ['Cmd', 'Shift']).getAllWindows().then(function(windows){
-  window.forEach(console.log);
-});
-```
-
-#### getWindowFrame
-`getWindowFrame( window ) ⇒ object`
-
+**getWindowFrame**  
+`getWindowFrame( window ) ⇒ object`  
 Return a window object with the `id` of the window and the `frame` field that
 contains the origin (top left) and size of the window.
 
@@ -106,35 +79,174 @@ z.bind('t', ['Cmd', 'Shift'])
 });
 ```
 
-#### getWindowOrigin
-`getWindowOrigin( window ) ⇒ object`
+**setWindowFrame**  
+`setWindowFrame( window )`  
+Set the window frame for the window identified by `id`.
 
-Return a window object with the `id` of the window and the `origin` field that
-contains the origin (top left corner) of the window.
+```node
+z.bind('t', ['Cmd', 'Shift'])
+.getFocusedWindow()
+.getWindowFrame()
+.SetWindowFrame(function(window){
+  window.frame.w /= 2;
+  return window;
+});
+```
+
+**getVisibleWindows**  
+`getVisibleWindows() ⇒ array`  
+Return an array containing an object for every visible window.
+
+```node
+z.bind('t', ['Cmd', 'Shift']).getVisibleWindows().then(function(windows){
+  window.forEach(console.log);
+});
+```
+
+**getAllWindows**  
+`getAllWindows() ⇒ array`  
+Return an array containing an object for all the windows.
+
+```node
+z.bind('t', ['Cmd', 'Shift']).getAllWindows().then(function(windows){
+  window.forEach(console.log);
+});
+```
+
+**getWindowTitle**  
+`getWindowTitle( window ) ⇒ object`  
+Return a window object with the `id` of the window and the `title`.
+
+```node
+z.bind('t', ['Cmd', 'Shift'])
+.getFocusedWindow()
+.getWindowTitle()
+.then(function(window){
+  console.log(window.title);
+});
+```
+
+**getWindowOrigin**  
+`getWindowOrigin( window ) ⇒ object`  
+Return a window object with the `id` of the window and the `origin` (top
+left corner) of the window.
 
 ```node
 z.bind('t', ['Cmd', 'Shift'])
 .getFocusedWindow()
 .getWindowOrigin()
 .then(function(window){
-  console.log(window.origin); // {x: 0, y: 0}
+  console.log(window.frame); // {x: 0, y: 0}
 });
 ```
 
-#### getWindowSize
-`getWindowSize( window ) ⇒ object`
-
-Return a window object with the `id` of the window and the `size` field that
-contains the size of the window.
+**setWindowOrigin**  
+`setWindowOrigin( window )`  
+Set the `origin` (top left corner) for a certain window `id`.
 
 ```node
 z.bind('t', ['Cmd', 'Shift'])
 .getFocusedWindow()
 .getWindowOrigin()
-.then(function(window){
-  console.log(window.origin); // {x: 0, y: 0}
+.setWindowOrigin(function(window){
+  window.frame.x += 100;
+  return window;
 });
 ```
+
+**getWindowSize**  
+`getWindowSize( window ) ⇒ object`  
+Return a window object with the `id` of the window and the `size`.
+
+```node
+z.bind('t', ['Cmd', 'Shift'])
+.getFocusedWindow()
+.getWindowSize()
+.then(function(window){
+  console.log(window.frame); // {w: 100, h: 100}
+});
+```
+
+**setWindowSize**  
+`setWindowSize( window )`  
+Set the window the `size` fot a certain window `id`.
+
+```node
+z.bind('t', ['Cmd', 'Shift'])
+.getFocusedWindow()
+.getWindowSize()
+.then(function(window){
+  window.frame.w += 200;
+  return window;
+});
+```
+
+**windowMaximize**  
+`windowMaximize( window )`  
+Maximize the window identified by `id`.
+
+```node
+z.bind('t', ['Cmd', 'Shift'])
+.getFocusedWindow()
+.windowMaximize();
+```
+
+**windowMinimize**  
+`windowMinimize( window )`  
+Minimize the window identified by `id`.
+
+```node
+z.bind('t', ['Cmd', 'Shift'])
+.getFocusedWindow()
+.windowMinimize();
+```
+
+**windowUnminimize**  
+`windowUnminimize( window )`  
+Unminimize the window identified by `id`.
+
+```node
+z.bind('t', ['Cmd', 'Shift'])
+.getFocusedWindow()
+.windowUnminimize();
+```
+
+**getAppFromWindow**  
+`getAppFromWindow( window ) ⇒ object`  
+Get the app object from the window identified by `id`.
+
+```node
+z.bind('t', ['Cmd', 'Shift'])
+.getFocusedWindow()
+.getAppFromWindow()
+.then(function(app){
+  console.log(app); //{ id: 1 }
+});
+```
+
+**getScreenFromWindow**  
+`getScreenFromWindow( window ) ⇒ object`  
+Get the screen object from the window identified by `id`.
+
+```node
+z.bind('t', ['Cmd', 'Shift'])
+.getFocusedWindow()
+.getScreenFromWindow()
+.then(function(screen){
+  console.log(screen); //{ id: 2 }
+});
+```
+
+**focusWindow{up, down, right, left}**  
+`focusWindow{up, down, right, left}( window )`  
+Focus the window identified by `id` to the {right, left, up, down}.
+
+```node
+z.bind('t', ['Cmd', 'Shift'])
+.getFocusedWindow()
+.focusWindowUp();
+```
+
 ## Tests
 Tests are written in [Mocha](http://visionmedia.github.io/mocha/). Simply run
 the test with:
