+++ package.json
@@ -7,7 +7,8 @@
   },
   "dependencies": {
     "when" : "2.3.0",
-    "node-uuid" : "1.4.0"
+    "node-uuid" : "1.4.0",
+    "lodash" : "1.3.1"
   },
   "devDependencies": {
     "mocha": "1.12.0"
